+++ netlify/functions/user.ts
@@ -2,11 +2,46 @@ import type { Context } from "@netlify/functions";
 import type { User } from "../../src/types";
 import { fetchGraphQL } from "./utils";
 
+const GETUSERSURL = "https://shopping-cart.hasura.app/api/rest/getusers/";
+
 export default async (req: Request, context: Context) => {
 	const { method } = req;
 	const { params } = context;
 
 	switch (method) {
+		case "GET":
+			try {
+				const resp: User[] = (
+					await (
+						await fetch(GETUSERSURL, {
+							headers: {
+								"Content-Type": "application/json",
+								"x-hasura-admin-secret":
+									process.env.VITE_HASURA_ADMIN_SECRET || "",
+							},
+						})
+					).json()
+				).user;
+
+				return new Response(JSON.stringify(resp || []), {
+					headers: {
+						"Content-Type": "application/json",
+					},
+					status: 200,
+				});
+			} catch (error) {
+				console.error("Error fetching users:", error);
+				return new Response(
+					JSON.stringify({ error: "Failed to fetch users" }),
+					{
+						headers: {
+							"Content-Type": "application/json",
+						},
+						status: 500,
+					},
+				);
+			}
+
 		case "POST": {
 			const username = decodeURI(params.username);
 			const variables = {
@@ -84,4 +119,4 @@ export default async (req: Request, context: Context) => {
 	});
 };
 
-export const config = { path: "/api/user/:username" };
+export const config = { path: ["/api/user/:username", "/api/user"] };
